move methods to existing extension class
